rename websocket app to ws (avoid naming conflicts with other packages)
