+++ mrmap/MrMap/asgi.py
@@ -10,7 +10,7 @@ os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'MrMap.settings')
 django_asgi_app = get_asgi_application()
 
 from channels.routing import ProtocolTypeRouter, URLRouter # noqa
-import websockets.routing # noqa
+import ws.routing # noqa
 from channels.security.websocket import AllowedHostsOriginValidator # noqa
 from channels.auth import AuthMiddlewareStack # noqa
 
@@ -21,7 +21,7 @@ application = ProtocolTypeRouter({
 
   "websocket": AllowedHostsOriginValidator(AuthMiddlewareStack(
       URLRouter(
-          websockets.routing.websocket_urlpatterns,
+          ws.routing.websocket_urlpatterns,
       )
   ),),
 })
