+++ mrmap/ws/static/websockets/js/websockets.js
@@ -6,6 +6,7 @@ function ws_connect(path, search) {
 
     ws_socket.onopen = function open() {
         // do nothing if the websocket is correctly opened
+        console.log('connected');
     };
 
     ws_socket.onclose = function(e) {
@@ -24,7 +25,6 @@ function ws_connect(path, search) {
 
 function handle_toast(){
     var ws_socket = ws_connect('/ws/toasts/', '')
-    console.log(ws_socket);
     ws_socket.onmessage = function message(event) {
         var json_data = JSON.parse(event.data);
         var toast_container = document.getElementById("id_toast_container");
